Check for InCombatLockDown() to decide which checkers can be used
In an hotfix CheckInteractDistance and IsItemInRange were made
protected in combat.

Blizzard in their incompetence somehow managed to not release that
in 10.2 as planned and didn't tell us about the plan to change it in 10.2.
And didn't tell us in advance of the hotfix releasing, nor could they
clarify whether the hotfix was intentionally for several hours after the
hotfix was released.