Fix spell checkers on Classic/SOD
Sspells with have a min range, e.g. a spell with an 8-30y range, cannot
be easily used of range checking, because if it isn't in range, that
could be <8 or >30. Thus we used a second checker to distinguish between
those two cases.

Previously only other spells would be consider, but in some cases
classes have no spells that can fill that gap.

Thus use an idea by kodewdle to fall-back to the interact list. All we
need is an interact distance that is bigger than min range and less than
the max range.

Fixes: #16