Workaround for unlearned spellbook spells in TWW (#30)
* Switching to `C_Spell.IsSpellInRange` for TWW

This replaces `C_SpellBook.IsSpellBookItemInRange` with `C_Spell.IsSpellInRange` in the TWW fallback, to ensure even unlearned spells can be used for range check. `IsSpellBookItemInRange` does not check unlearned spells.
To make this work, the spell index used and returned from `findSpellIdx` is now the `spellID` in retail, but still the spellbook index for classic/era.