Hero Talent spell overrides don't check range.
Living Flame, Smite and Lightning Bolt can be temporarily replaced by an override spell, rendering the spell useless for range checks.
This adds "fallback" spells with a higher priority which will be picked instead.
Emerald Blossom is available and unmodified for Pres and Aug Evoker.
Mind Blast is available for Disc Priest, where the override happens and no SW:P is available.
Earth Shock and Elemental blast share the same choice node and act as a replacement for Lightning Bold in this case for Elemental Shaman.